% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albumin.R
\name{albumin_charge}
\alias{albumin_charge}
\title{Estimate graphical albumin charge}
\usage{
albumin_charge(albumin)
}
\arguments{
\item{albumin}{Albumin concentration in g/dL.}
}
\value{
Estimated albumin charge in mEq/L.
}
\description{
This simplified estimate is used only for the Gamblegram visualization:
\code{2.8 * albumin_g_dl}.
}
