% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{anion_components}
\alias{anion_components}
\title{Build anion components}
\usage{
anion_components(chloride, bicarbonate, lactate, albumin, residual_anions = 0)
}
\arguments{
\item{chloride}{Chloride concentration in mmol/L.}

\item{bicarbonate}{Bicarbonate or total CO2 concentration in mmol/L.}

\item{lactate}{Lactate concentration in mmol/L.}

\item{albumin}{Albumin concentration in g/dL.}

\item{residual_anions}{Illustrative residual anion compartment in mEq/L.}
}
\value{
A named numeric vector of anion charges.
}
\description{
Build anion components
}
