% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_card.R
\name{deal_card}
\alias{deal_card}
\title{Deal the top card from a deck}
\usage{
deal_card(cards = cards_df)
}
\arguments{
\item{cards}{A two dimensional object with 52 rows and three columns, including all data within a deck of cards;faces, suits and values of each of the 52 cards in a typical deck.}
}
\value{
The top card in the deck will be dealt, providing the face, suite and value of the card.
}
\description{
A function that will deal the top card from the deck or cards dataframe.
}
\examples{
dealt_cards <- deal_card(cards = cards_df)

deal_card(cards = cards_df)
}
