% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{get_component_labels}
\alias{get_component_labels}
\title{Get translated component labels}
\usage{
get_component_labels(language = "es")
}
\arguments{
\item{language}{Language code. Supported values are \code{"es"} and \code{"en"}.}
}
\value{
A named character vector.
}
\description{
Get translated component labels
}
