% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{get_plot_labels}
\alias{get_plot_labels}
\title{Get translated Gamblegram plot labels}
\usage{
get_plot_labels(language = "es")
}
\arguments{
\item{language}{Language code. Supported values are \code{"es"} and \code{"en"}.}
}
\value{
A list containing translated component labels, axis labels, and
vertical-axis label.
}
\description{
Get translated Gamblegram plot labels
}
