% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{implied.probability}
\alias{implied.probability}
\title{Calculate implied probability from odds.}
\usage{
implied.probability(odds)
}
\arguments{
\item{odds}{A vector of either fractional or decimal odds.}
}
\description{
Calculate implied probability from odds.
}
\details{
Implied probability is the reciprocal of the decimal odds.

Decimcal odds of 2.0 are equivalent to an implied probability of 50%.
}

