% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_payout.R
\name{line_payout}
\alias{line_payout}
\title{Calculate payout for a given line with a wager}
\usage{
line_payout(line, wager)
}
\arguments{
\item{line}{Money line}

\item{wager}{Wager}
}
\value{
Total payout (wager + winnings)
}
\description{
Calculate payout for a given line with a wager
}
