% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_to_perc.R
\name{line_to_perc}
\alias{line_to_perc}
\title{Convert money line into a decimal win percentage}
\usage{
line_to_perc(line)
}
\arguments{
\item{line}{Money line}
}
\value{
Win percentage as a decimal
}
\description{
Convert money line into a decimal win percentage
}
\examples{
lines <- c(100, -100, 150)
line_to_perc(lines)
}
