% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gamblegram_inputs.R
\name{mod_gamblegram_inputs_server}
\alias{mod_gamblegram_inputs_server}
\title{Gamblegram inputs module server}
\usage{
mod_gamblegram_inputs_server(id)
}
\arguments{
\item{id}{Shiny module identifier.}
}
\value{
A list containing reactive expressions for language and chemistry.
}
\description{
Gamblegram inputs module server
}
