% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gamblegram_inputs.R
\name{mod_gamblegram_inputs_ui}
\alias{mod_gamblegram_inputs_ui}
\title{Gamblegram inputs module UI}
\usage{
mod_gamblegram_inputs_ui(id)
}
\arguments{
\item{id}{Shiny module identifier.}
}
\value{
A Shiny tag list.
}
\description{
Gamblegram inputs module UI
}
