% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gamblegram_output.R
\name{mod_gamblegram_output_server}
\alias{mod_gamblegram_output_server}
\title{Gamblegram output module server}
\usage{
mod_gamblegram_output_server(id, chemistry, language)
}
\arguments{
\item{id}{Shiny module identifier.}

\item{chemistry}{Reactive expression containing chemistry values.}

\item{language}{Reactive expression containing \code{"es"} or \code{"en"}.}
}
\value{
Invisibly returns a list of internal reactive expressions, primarily
for testing.
}
\description{
Gamblegram output module server
}
