% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gamblegram_output.R
\name{mod_gamblegram_output_ui}
\alias{mod_gamblegram_output_ui}
\title{Gamblegram output module UI}
\usage{
mod_gamblegram_output_ui(id)
}
\arguments{
\item{id}{Shiny module identifier.}
}
\value{
A Shiny tag list.
}
\description{
Gamblegram output module UI
}
