% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddschecker.R
\name{oddschecker}
\alias{oddschecker}
\title{Scrape odds off http://www.oddschecker.com/.}
\usage{
oddschecker(event)
}
\arguments{
\item{event}{Event specification.}
}
\value{
Table of odds quoted by various bookmakers for each of the contenders.
}
\description{
Scrape odds off http://www.oddschecker.com/.
}
\examples{
oddschecker("politics/us-politics/us-presidential-election-2016/winner")
oddschecker("horse-racing/carlisle/18:05/winner")
}

