% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlay_to_line.R
\name{parlay_to_line}
\alias{parlay_to_line}
\title{Parlays}
\usage{
parlay_to_line(...)
}
\arguments{
\item{...}{Money lines}
}
\value{
Money line for the combined bets assuming independence
}
\description{
Parlays
}
