% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlay_to_perc.R
\name{parlay_to_perc}
\alias{parlay_to_perc}
\title{Convert parlay to win percentage}
\usage{
parlay_to_perc(...)
}
\arguments{
\item{...}{Money lines}
}
\value{
A single percentage representing the win percentage
}
\description{
Convert parlay to win percentage
}
