% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_to_line.R
\name{perc_to_line}
\alias{perc_to_line}
\title{Convert decimal percentage into a money line}
\usage{
perc_to_line(perc)
}
\arguments{
\item{perc}{Win percentage as a decimal}
}
\value{
Money line
}
\description{
Convert decimal percentage into a money line
}
