% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_blackjack.R
\name{play_blackjack}
\alias{play_blackjack}
\title{Step up and face the dealer in Blackjack!}
\usage{
play_blackjack(bet = 10, currency = "£")
}
\arguments{
\item{bet}{A positive integer representing the amount you're willing to gamble (and probably lose). Default is \code{10}.}

\item{currency}{A string representing the currency the user wishes to play with. Defaults to GBP (\verb{£}), but feel free to flex in dollars, euros, or even Monopoly money.}
}
\value{
A log of the game, showing the cards drawn and the final outcome.
}
\description{
Blackjack is a classic casino game where the goal is to beat the dealer by getting as close to 21 as possible without going over.
\subsection{House Rules:}{
\itemize{
\item Aces can count as \strong{1 or 11}.
\item Face cards (Jack, Queen and King) are worth \strong{10}.
\item You can \strong{stick} (stay with your current hand) or \strong{twist} (hit for another card).
\item Dealer must \strong{twist on 16 or lower}.
\item If a players first two cards make \strong{21}, and contain an Ace it's Blackjack! (Automatic win unless dealer also has it, resulting in a draw).
}
}
}
\examples{
# Playing it safe with the default £10 stake
GambleR::play_blackjack()

# Big-doggin' it
GambleR::play_blackjack(bet = 500,
                        currency = "£")
}
