% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_roulette.R
\name{play_roulette}
\alias{play_roulette}
\title{Spin the wheel at GambleR's roulette table!}
\usage{
play_roulette(prediction = "Black", bet = 10, roulette_type = "European")
}
\arguments{
\item{prediction}{A string stating which colour you wish to bet on. Choose from "Red", "Black", or "Green". Red and Black offer even-money odds (2:1), while Green pays out big at 35:1 (European) or 38:2 (American) because, well, it’s called \emph{gambling} for a reason.}

\item{bet}{A positive number representing how much you’re willing to risk in imaginary casino currency. The default currency is GBP(£) when playing on a European table, but if you're playing on an \code{American} table, it'll be in USD($).}

\item{roulette_type}{The type of roulette table you want to play on. The default is \code{European}, where there's only one Green zero. Feeling less-bold? Play the \code{American} table for more favourable odds on Green!}
}
\value{
The function returns the result of the spin, announcing where the ball landed and whether you won big, won small, or should probably rethink your life choices.
}
\description{
Step right up and test your luck at the GambleR roulette table! Will you walk away a legend, or will the house (which always wins) claim yet another victim? This function lets you place a simple bet on Red, Black, or Green—without the risk of re-mortgaging your house, putting it all on Black, and explaining your life choices to your now ex-spouse.
}
\examples{
# A cautious gambler playing it safe:
GambleR::roulette(prediction = "Black",
                  bet = 10,
                  roulette_type = "European")

# Going for the big Green payout:
GambleR::roulette(prediction = "Green",
                  bet = 100,
                  roulette_type = "American")
}
