% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_dice.R
\name{roll_dice}
\alias{roll_dice}
\title{Roll the dice and test your luck!}
\usage{
roll_dice(number_of_die = 1)
}
\arguments{
\item{number_of_die}{A positive integer, stating the number of dice to roll. Default value is 1.}
}
\value{
A numeric value representing the total sum of all dice rolled.
}
\description{
Simulates rolling one or more dice. Will you roll a lucky seven, or just a handful of disappointment?
}
\examples{
# Roll a single die
roll_dice()

# Roll two dice (classic style)
roll_dice(2)
}
