% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_blackjack}
\alias{score_blackjack}
\title{Score Blackjack}
\usage{
score_blackjack(x, y = NULL)
}
\arguments{
\item{game}{data.frame(1) or numeric(n). If a data.frame is provided, it must contain columns card_value and player_id.
The data.frame should be in long format where a row represents one card in a players hand. If a numeric vector is
provided it should just be the card value. See Details.}
}
\value{
list().
player_score = numeric. the players score.
soft_value = logical. The player has a soft ace (ace treated as 11).
}
\description{
Score Blackjack
}
\details{
The card value should be a three digit number code where the first digit represents the suit (1, 2, 3, or 4) and the
last two digits represent the card code (02-14). Values 11-13 represent the face cards (Jack, Queen, King) while 14
represents the Ace. Face cards are automatically converted to be worth 10 points. The Ace is counted as a 1 if
treating it as an 11 would result in the player busting (exceeding a score of 21), otherwise it is treated as an 11.
}
