% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{Re-shuffle a deck of cards randomly and fairly}
\usage{
shuffle(cards = cards_df)
}
\arguments{
\item{cards}{A two dimensional object with 52 rows and three columns, including all data within a deck of cards;faces, suits and values of each of the 52 cards in a typical deck.}
}
\value{
The deck of cards will be randomly re-shuffled.
}
\description{
Re-shuffle a deck of cards randomly and fairly
}
\examples{
cards_df <- shuffle(cards = cards_df)
shuffle(cards = cards_df)
}
