% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_cards.R
\name{shuffle_cards}
\alias{shuffle_cards}
\title{Re-shuffle a deck of virtual playing cards}
\usage{
shuffle_cards(cards = NULL)
}
\arguments{
\item{cards}{A data frame containing a standard deck of 52 playing cards.
If no deck is provided, a fresh one will be generated.}
}
\value{
A shuffled deck of playing cards as a data frame.
}
\description{
Randomly reorders a deck of 52 playing cards, ensuring a fair shuffle.
The shuffled deck object can then be used in other functions such as \code{GambleR::deal_card()}.
}
\examples{
# Shuffle a fresh deck
shuffled_deck <- GambleR::shuffle_cards()

# Shuffle an existing deck
shuffled_deck <- GambleR::shuffle_cards(cards = deck_of_cards)
}
