% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_blackjack.R
\name{simulate_blackjack}
\alias{simulate_blackjack}
\title{Simulate Blackjack}
\usage{
simulate_blackjack(n_players, n_hands, n_decks = 1, count_method = "hi-lo")
}
\arguments{
\item{n_players}{numeric(1). The number of players (not including the dealer) who are playing in each hand.}

\item{n_hands}{numeric(1). The number of hands to simulate.}

\item{n_decks}{numeric(1). The number of decks used for dealing.}

\item{count_method}{character(1). The method used to count cards. The count ~will eventually~ impact the hitting
strategy of the players. One of c("hi-lo", "omega2", "griffin_ultimate"). Default is "hi-lo". See Details.}
}
\value{
A tibble containing the following columns:
\item{shuffle_id}{The number of times the deck has been shuffled since the start.}
\item{hand_id}{The unique hand ID.}
\item{player_id}{The ID/seat for each player (101 represents the dealer).}
\item{card_number}{The number of cards a player has been dealt within a given hand.}
\item{card_id}{The dealt card's ID. The 100s digit represents suit (1-4) while the 10s and 1s digits
represents the face value (2-14: 11-13 represent Jack, Queen, King respectively and 14 represents the Ace)}
\item{card_value}{The dealt card's value. Always 1 for an Ace even though an Ace can also be worth 11.}
\item{running_score}{A player's running score within the hand (First two dealt cards have the same score).}
\item{deck_count}{The knowable count of the deck (not true count since we don't count the first burned card.}
}
\description{
Simulates a large number of hands of blackjack efficiently.
}
\details{
The simulation closely mimics how blackjack is dealt in a casino. This includes burning one card after shuffling and
not reshuffling until the deck gets below a pre-determined point. The dealer's hit strategy is hit on a soft
seventeen or less. The players' hit strategy is currently the same as the dealer's. In the future, there will be an
option to select the hit strategy using the \code{count_method} and to vary the hit strategy for each individual player.
\code{count_method} options are pulled from \href{https://www.blackjackreview.com/wp/encyclopedia/card-counting-system-comparisons/}{here}.
}
