% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slot_machine.R
\name{slot_machine}
\alias{slot_machine}
\title{Test your luck on the gambleR slot machine}
\usage{
slot_machine()
}
\value{
The slot machine's wheel will be spun, producing a combination of three symbols as well as the amount won in $ (with any luck). The result of the spin will be printed in the console.
}
\description{
Test your luck on the gambleR slot machine
}
\examples{
result <- slot_machine()
slot_machine()
}
