% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{to.fractional}
\alias{to.fractional}
\title{Convert decimal odds to fractional odds.}
\usage{
to.fractional(decimal, ...)
}
\arguments{
\item{decimal}{A numeric vector of decimal odds.}
}
\value{
A character vector of fractional odds.
}
\description{
Convert decimal odds to fractional odds.
}
\examples{
to.fractional(c(3, 2.5, 1.25))
}

