% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourdefrance.R
\docType{data}
\name{tourdefrance}
\alias{tourdefrance}
\title{Tour de France Odds}
\format{A data frame:
\describe{
  \item{rider}{Name of rider.}
  \item{odds}{Decimal odds.}
}}
\source{
Data gathered from https://sport.netbet.co.uk/cycling/tour-de-france/ on 16 July 2016.
}
\usage{
tourdefrance
}
\description{
Tour de France Odds
}
\keyword{datasets}

