% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{validate_language}
\alias{validate_language}
\title{Validate a supported language}
\usage{
validate_language(language)
}
\arguments{
\item{language}{Language code.}
}
\value{
The validated language code.
}
\description{
Validate a supported language
}
