% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_card.R
\name{deal_card}
\alias{deal_card}
\title{Deal one or more cards from a deck of playing cards}
\usage{
deal_card(cards = NULL, n_cards = 1, verbose = TRUE)
}
\arguments{
\item{cards}{A data frame representing a deck of cards. If not provided, a freshly shuffled deck derived from \code{GambleR::shuffle_cards()} will be used each time the function is called with no \code{cards} object provided.}

\item{n_cards}{The number of cards to deal. Default is 1.}

\item{verbose}{A logical input (TRUE or FALSE), to determine whether the function prints the output of the cards. This argument is TRUE by default.}
}
\value{
A list with two elements:
\itemize{
\item \code{dealt}: The card(s) dealt.
\item \code{remaining_deck}: The updated deck with dealt cards removed.
}
}
\description{
Deals the top card(s) from a given deck, returning both the dealt card(s) and the updated deck.
}
\examples{
# Deal a single card from a fresh deck
dealt_card <- GambleR::deal_card()
dealt_card <- dealt_card$dealt
remaining_deck <- dealt_card$remaining_deck

# Deal two cards from an existing deck
dealt_cards <- GambleR::deal_card(cards = remaining_deck,
                                  n_cards = 2)
dealt_cards <- dealt_cards$dealt
updated_deck <- dealt_cards$remaining_deck
}
