% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_blackjack.R
\name{play_blackjack}
\alias{play_blackjack}
\title{Step up and face the dealer in Blackjack!}
\usage{
play_blackjack(bet = 10, currency = "£")
}
\arguments{
\item{bet}{A positive integer representing the amount you're willing to gamble (and probably lose). Default is \code{10}.}

\item{currency}{A string representing the currency the user wishes to play with. Defaults to GBP (\verb{£}), but feel free to flex in dollars, euros, or even Monopoly money.}
}
\value{
A log of the game, showing the cards drawn and the final outcome.
}
\description{
Step up and face the dealer in Blackjack!
}
\examples{
# Playing it safe with the default £10 stake
play_blackjack()

# Big-dogging it
play_blackjack(bet = 500,
               currency = "£")
}
