% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll a pair of dice}
\usage{
roll()
}
\value{
A random score between 2 - 12, calculated by adding the score of two virtual dice together.
}
\description{
A function that simulates the rolling of a pair of dice (or die).
}
\examples{
score <- roll()
roll()
}
