% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roulette.R
\name{roulette}
\alias{roulette}
\title{Spin the wheel at gambleR's roulette table}
\usage{
roulette(prediction = prediction, bet = bet)
}
\arguments{
\item{prediction}{A string stating which colour the gambler wishes to place their stake on. The colours are "Red", "Black" or "Green". The odds that the ball will land on "Red" or "Black" are 2/1, and the odds for "Green" are 35/1.}

\item{bet}{Any positive number, stating how much the gambler would like to bet on their prediction.}
}
\value{
The function will return the result of the spin, stating which colour the ball has landed on as well as any prize won, if any.
}
\description{
Spin the wheel and try your luck at guessing what colour the virtual ball will land on. This function resembles a very basic roulette table, a very fun way to gamble away any spare funds.
}
\examples{
result <- roulette(prediction, "Red", bet = 10)

roulette(prediction = "Black", bet = 10)
}
