% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{to.decimal}
\alias{to.decimal}
\title{Convert fractional odds to decimal odds.}
\usage{
to.decimal(fractional)
}
\arguments{
\item{fractional}{A character vector of fractional odds.}
}
\value{
A numeric vector of decimal odds.
}
\description{
Convert fractional odds to decimal odds.
}
\examples{
to.decimal(c("2/1", "5/3", "1/4"))
}

